Change summary total fields from int64 to int32
Updated all relevant structs, protobuf definitions, and OpenAPI specs to use int32 instead of int64 for 'total' fields in summary-related types. This ensures consistency across the API, domain models, and storage layers.

+++ api/handler/v1beta1/approval.go
@@ -168,7 +168,7 @@ func (s *GRPCServer) GenerateListUserApprovalsSummaries(ctx context.Context, req
 
 	return &guardianv1beta1.GenerateListUserApprovalsSummariesResponse{
 		Summaries: summaries,
-		Total:     int64(len(summaries)),
+		Total:     int32(len(summaries)),
 	}, nil
 }
